Reoganize remote client code to make room for non-ssh remote projects
