+++ crates/call/src/call_impl/room.rs
@@ -1161,7 +1161,7 @@ impl Room {
         let request = self.client.request(proto::ShareProject {
             room_id: self.id(),
             worktrees: project.read(cx).worktree_metadata_protos(cx),
-            is_ssh_project: project.read(cx).is_via_ssh(),
+            is_ssh_project: project.read(cx).is_via_remote_server(),
         });
 
         cx.spawn(async move |this, cx| {
