+++ crates/remote/src/remote.rs
@@ -1,9 +1,11 @@
 pub mod json_log;
 pub mod protocol;
 pub mod proxy;
-pub mod ssh_session;
+pub mod remote_client;
+mod transport;
 
-pub use ssh_session::{
-    ConnectionState, SshClientDelegate, SshConnectionOptions, SshInfo, SshPlatform,
-    SshRemoteClient, SshRemoteEvent,
+pub use remote_client::{
+    ConnectionIdentifier, ConnectionState, RemoteClient, RemoteClientDelegate, RemoteClientEvent,
+    RemotePlatform,
 };
+pub use transport::ssh::{SshArgs, SshConnectionOptions, SshInfo, SshPortForwardOption};
